+++ plugins/destination/clickhouse/typeconv/arrow/types/nullable.go
@@ -2,7 +2,7 @@ package types
 
 import (
 	"github.com/ClickHouse/clickhouse-go/v2/lib/column"
-	"github.com/apache/arrow/go/v15/arrow"
+	"github.com/apache/arrow/go/v16/arrow"
 )
 
 func nullableType(name string, col *column.Nullable) (*arrow.Field, error) {