+++ plugins/destination/clickhouse/typeconv/arrow/values/record.go
@@ -3,9 +3,9 @@ package values
 import (
 	"fmt"
 
-	"github.com/apache/arrow/go/v15/arrow"
-	"github.com/apache/arrow/go/v15/arrow/array"
-	"github.com/apache/arrow/go/v15/arrow/memory"
+	"github.com/apache/arrow/go/v16/arrow"
+	"github.com/apache/arrow/go/v16/arrow/array"
+	"github.com/apache/arrow/go/v16/arrow/memory"
 )
 
 func Record(sc *arrow.Schema, data []any) (arrow.Record, error) {