+++ plugins/destination/s3/client/write.go
@@ -7,18 +7,40 @@ import (
 	"regexp"
 	"time"
 
-	"github.com/apache/arrow/go/v15/arrow"
-	"github.com/apache/arrow/go/v15/arrow/array"
-	"github.com/apache/arrow/go/v15/arrow/memory"
+	"github.com/apache/arrow/go/v16/arrow"
+	"github.com/apache/arrow/go/v16/arrow/array"
+	"github.com/apache/arrow/go/v16/arrow/memory"
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/s3"
 	"github.com/cloudquery/filetypes/v4"
 	"github.com/cloudquery/plugin-sdk/v4/message"
+	"github.com/cloudquery/plugin-sdk/v4/schema"
 	"github.com/cloudquery/plugin-sdk/v4/types"
 	"github.com/google/uuid"
 )
 
 var reInvalidJSONKey = regexp.MustCompile(`\W`)
+
+func (c *Client) createObject(ctx context.Context, table *schema.Table, objKey string) (*filetypes.Stream, error) {
+	s, err := c.Client.StartStream(table, func(r io.Reader) error {
+		params := &s3.PutObjectInput{
+			Bucket:      aws.String(c.spec.Bucket),
+			Key:         aws.String(objKey),
+			Body:        r,
+			ContentType: aws.String(c.spec.GetContentType()),
+		}
+
+		sseConfiguration := c.spec.ServerSideEncryptionConfiguration
+		if sseConfiguration != nil {
+			params.SSEKMSKeyId = &sseConfiguration.SSEKMSKeyId
+			params.ServerSideEncryption = sseConfiguration.ServerSideEncryption
+		}
+
+		_, err := c.uploader.Upload(ctx, params)
+		return err
+	})
+	return s, err
+}
 
 func (c *Client) WriteTable(ctx context.Context, msgs <-chan *message.WriteInsert) error {
 	var s *filetypes.Stream
@@ -26,27 +48,16 @@ func (c *Client) WriteTable(ctx context.Context, msgs <-chan *message.WriteInser
 	for msg := range msgs {
 		if s == nil {
 			table := msg.GetTable()
-
 			objKey := c.spec.ReplacePathVariables(table.Name, uuid.NewString(), time.Now().UTC(), c.syncID)
+			// if object was already initialized, use the same key
+			// We don't need any locking here because all messages for the same table are processed sequentially
+			if val, ok := c.initializedTables[table.Name]; ok {
+				objKey = val
+				delete(c.initializedTables, table.Name)
+			}
 
 			var err error
-			s, err = c.Client.StartStream(table, func(r io.Reader) error {
-				params := &s3.PutObjectInput{
-					Bucket:      aws.String(c.spec.Bucket),
-					Key:         aws.String(objKey),
-					Body:        r,
-					ContentType: aws.String(c.spec.GetContentType()),
-				}
-
-				sseConfiguration := c.spec.ServerSideEncryptionConfiguration
-				if sseConfiguration != nil {
-					params.SSEKMSKeyId = &sseConfiguration.SSEKMSKeyId
-					params.ServerSideEncryption = sseConfiguration.ServerSideEncryption
-				}
-
-				_, err := c.uploader.Upload(ctx, params)
-				return err
-			})
+			s, err = c.createObject(ctx, table, objKey)
 			if err != nil {
 				return err
 			}
@@ -72,6 +83,27 @@ func (c *Client) WriteTable(ctx context.Context, msgs <-chan *message.WriteInser
 
 func (c *Client) Write(ctx context.Context, msgs <-chan message.WriteMessage) error {
 	return c.writer.Write(ctx, msgs)
+}
+
+func (c *Client) MigrateTable(ctx context.Context, ch <-chan *message.WriteMigrateTable) error {
+	for msg := range ch {
+		if !c.spec.GenerateEmptyObjects {
+			continue
+		}
+		table := msg.GetTable()
+		objKey := c.spec.ReplacePathVariables(table.Name, uuid.NewString(), time.Now().UTC(), c.syncID)
+		// We don't need any locking here because all messages for the same table are processed sequentially
+		c.initializedTables[table.Name] = objKey
+		s, err := c.createObject(ctx, table, objKey)
+		if err != nil {
+			return err
+		}
+		err = s.Finish()
+		if err != nil {
+			return err
+		}
+	}
+	return nil
 }
 
 // sanitizeRecordJSONKeys replaces all invalid characters in JSON keys with underscores. This is required