+++ plugins/source/aws/tools/tool.go
@@ -1,9 +1,0 @@
-//go:build tools
-// +build tools
-
-package tools
-
-import (
-	_ "github.com/golang/mock/mockgen"
-	_ "github.com/golang/mock/mockgen/model"
-)