+++ plugins/source/gitlab/client/services/projects.go
@@ -1,8 +1,0 @@
-package services
-
-import "github.com/xanzy/go-gitlab"
-
-type ProjectsClient interface {
-	ListProjects(opt *gitlab.ListProjectsOptions, options ...gitlab.RequestOptionFunc) ([]*gitlab.Project, *gitlab.Response, error)
-	GetProject(pid any, opt *gitlab.GetProjectOptions, options ...gitlab.RequestOptionFunc) (*gitlab.Project, *gitlab.Response, error)
-}